TPM2 ACPI table support
Add a TPM2 ACPI table if a TPM 2 is used in the backend.
Also add an SSDT for the TPM 2.

Rename tpm_find() to tpm_get_version() and have this function
return the version of the TPM found, TPMVersion_Unspec if
no TPM is found. Use the version number to build version
specific ACPI tables.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>

---S

v3->v4:
   - added .hex.generated files
   - SSDT's for TPM v1.2 and TPM 2 now use common code via #include
     (enabling IRQ for TPM v1.2)

v2->v3:
   - followed Igor's suggestion of a default case with an assert()
   - added an SSDT for TPM2; it will be a bit different than TPM1.2's
     SSDT once we add PPI to it