Support Physical Presence Interface Spec
For automated management of a TPM device, implement the TCG Physical Presence
Interface Specification that allows a root user on Linux (for example) to set
an opcode for a sequence of TPM operations that the BIOS is supposed to execute
upon reboot of the physical or virtual machine. A sequence of operations may for
example involve giving up ownership of the TPM and activating and enabling the
device.

The sequences of operations are defined in table 2 in the specs to be found
at the following link:

http://www.trustedcomputinggroup.org/resources/tcg_physical_presence_interface_specification

As an example, in recent versions of Linux the opcode (5) can be set as
follows:

cd /sys/devices/pnp0/00\:04/ppi

echo 5 > request

This ACPI implementation assumes that the TPM interface has been extended
with vendor-specific memory locations that are not reset to zero upon a
machine reset but can hold the opcode that ACPI writes into it.
The BIOS indicates its support for this extensions by writing a 4-byte
signature 'TCG_MAGIC' at the beginning of the memory. ACPI code expects to
find this signature there.

The supported opcodes are 1-11, 14, and 21-22. (see table 2 in spec)
Also '0' is supported to 'clear' an intention.


Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Cc: Michael Tsirkin <mst@redhat.com>
Cc: Kevin O'Connor <kevin@koconnor.net>

---S

v3->v4:
  - consolidated code following Igor's suggestions
  - marked functions as Serialized
  - added .hex.generated files