Introduce RAM location in vendor specific area in TIS
Introduce RAM locations in the vendor specific area in the TIS. These
locations will survive a reset and will be part of the state written
during a suspend. Their puspose is to support the physical presence
interface where the OS (ACPI) and the firmware (SeaBIOS) use these RAM
locations to exchange data.

Only locality 0 is used, leaving localities 1-4 available for other extensions.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>