tpm: Support TPM for ppc64 using CRQ based interface
Implement support for TPM on ppc64 by implementing the vTPM CRQ
interface (following PAPR) as a frontend.

Have the frontend call the existing passthrough TPM device backend
using data structures from the TPM TIS hardware interface emulation
(tpm_tis.c). The TIS implements a superset of functionality and the
vTPM CRQ interface uses part of it, such as only 1 of the 5 localities.

The Linux vTPM driver for ppc64 works with this emulation.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>